// packages/datadog-api-client-v1/models/ApplicationKey.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
* An application key with its associated metadata.
*/
export declare class ApplicationKey {
    /**
    * Hash of an application key.
    */
    'hash'?: string;
    /**
    * Name of an application key.
    */
    'name'?: string;
    /**
    * Owner of an application key.
    */
    'owner'?: string;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): ApplicationKey;
    static serialize(data: ApplicationKey): {
        [key: string]: any;
    };
    constructor();
}