// packages/datadog-api-client-v1/models/Organization.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OrganizationBilling } from './OrganizationBilling';
import { OrganizationSettings } from './OrganizationSettings';
import { OrganizationSubscription } from './OrganizationSubscription';
/**
* Create, edit, and manage organizations.
*/
export declare class Organization {
    'billing'?: OrganizationBilling;
    /**
    * Date of the organization creation.
    */
    'created'?: string;
    /**
    * Description of the organization.
    */
    'description'?: string;
    /**
    * The name of the new child-organization, limited to 32 characters.
    */
    'name'?: string;
    /**
    * The `public_id` of the organization you are operating within.
    */
    'publicId'?: string;
    'settings'?: OrganizationSettings;
    'subscription'?: OrganizationSubscription;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): Organization;
    static serialize(data: Organization): {
        [key: string]: any;
    };
    constructor();
}