// packages/datadog-api-client-v1/models/User.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AccessRole } from './AccessRole';
/**
* Create, edit, and disable users.
*/
export declare class User {
    'accessRole'?: AccessRole;
    /**
    * The new disabled status of the user.
    */
    'disabled'?: boolean;
    /**
    * The new email of the user.
    */
    'email'?: string;
    /**
    * The user handle, must be a valid email.
    */
    'handle'?: string;
    /**
    * Gravatar icon associated to the user.
    */
    'icon'?: string;
    /**
    * The name of the user.
    */
    'name'?: string;
    /**
    * Whether or not the user logged in Datadog at least once.
    */
    'verified'?: boolean;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): User;
    static serialize(data: User): {
        [key: string]: any;
    };
    constructor();
}